backend/requirements.txt
aiohttp==3.7.4.post0
aiomysql==0.0.21
aioredis==1.3.1
fastapi==0.63.0
PyYAML==5.4.1
requests==2.25.1
SQLAlchemy==1.4.5
starlette==0.13.6
ujson==4.0.2
uvicorn==0.13.4
uvloop==0.15.2
